+++ kapacitor/client.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"net/http"
 	"regexp"
+	"strings"
 
 	"github.com/influxdata/chronograf"
 	"github.com/influxdata/chronograf/id"
@@ -69,7 +70,7 @@ type Task struct {
 	TICKScript chronograf.TICKScript // TICKScript is the running script
 }
 
-var reTaskName = regexp.MustCompile(`[\r\n]*var[ \t]+name[ \t]+=[ \t]+'([^']+)'`)
+var reTaskName = regexp.MustCompile(`[\r\n]*var[ \t]+name[ \t]+=[ \t]+'([^\n]+)'[ \r\t]*\n`)
 
 // NewTask creates a task from a kapacitor client task
 func NewTask(task *client.Task) *Task {
@@ -90,7 +91,7 @@ func NewTask(task *client.Task) *Task {
 	if rule.Name == "" {
 		// try to parse Name from a line such as: `var name = 'Rule Name'
 		if matches := reTaskName.FindStringSubmatch(task.TICKscript); matches != nil {
-			rule.Name = matches[1]
+			rule.Name = strings.ReplaceAll(strings.ReplaceAll(matches[1], "\\'", "'"), "\\\\", "\\")
 		} else {
 			rule.Name = task.ID
 		}