+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 
 1. [#5831](https://github.com/influxdata/chronograf/pull/5831): Add download button on query management page.
+1. [#5836](https://github.com/influxdata/chronograf/pull/5836): Allow to rename TICKscript.
 
 ### Bug Fixes